add(update) worklist web ui
